rename EditcarForm to EditCar and move the form from the old EditCarForm to a new component calles EditCarForm
