+++ src/Router.js
@@ -10,7 +10,7 @@ import Welcome from './components/Welcome';
 import Navbar from './containers/Navbar';
 import Appointments from './containers/Appointments';
 import CreateCar from './containers/CreateCar';
-import EditCarForm from './containers/EditCarForm';
+import EditCar from './containers/EditCar';
 
 const Router = ({ authToken }) => {
   if (authToken === '') {
@@ -33,7 +33,7 @@ const Router = ({ authToken }) => {
       <Navbar />
       <Switch>
         <Route exact path="/create_car" component={CreateCar} />
-        <Route exact path="/edit_car" component={EditCarForm} />
+        <Route exact path="/edit_car" component={EditCar} />
         <Route exact path="/" component={Home} />
         <Route exact path="/SignUp" component={SignUp} />
         <Route exact path="/login" component={Login} />
